UCS/RCACHE: fix .h file conflicts
Move rcache definition into separate internal .h file
